Call callbacks in predictable order
If a directory structure is modified, it is useful for the user of FSSM
to receive the callbacks in predictable order.

For example, let's consider an application that monitors files in a
working directory and attempts to keep the directory contents
synchronized to a remote host with SCP. If a subdirectory with a file
gets deleted in the working dir, it is useful for the app to know that
the file gets deleted first and the subdirectory second. This way, in
the remote host, the app can remove the file first before attempting to
remove the (non-empty) subdirectory.