Add option to include directories to be monitored
When monitoring the contents of a directory, if option :directories
is enabled, FSSM calls callbacks upon directory operations.

The interface of FSSM.monitor changes: if the last argument is a hash,
it is presumed to be the options hash and is passed to the monitor
object. This is a minimal API change, which hopefully does not break
existing code.

The options hash belongs to the monitor object, but the monitor object
passes options to its components of which behavior depend on options.