Pass file type argument to callbacks if :directories
If :directories option is enabled, callbacks receive a third argument
specifying the type of the file affected by the operation causing the
callback. The file types are either :file or :directory.

If :directories option is disabled, no third argument is passed to the
callbacks; the monitored files are all non-directories anyway.

With file creation and modification events, you can use
`File.directory?` to determine the file type. However, this is not
possible with file deletion, since the file does not exist anymore. This
is the motivation for the change.