Speed up monitor spec with a count down latch
Instead of sleeping a fixed amount of time before checking the
monitoring results, we use a count down latch to signal to the main
thread when expected number of events has occurred. This makes the test
run twice faster than originally.

Unfortunately, we still need a small period (1 s) to wait for the
monitor to start. Otherwise, the events we generate in a test might have
happened before the monitor is up and running.

Tested with
- JRuby 1.5.6,
- Rubinius 1.1.1,
- MRI 1.8.7-p302, and
- MRI 1.9.2-p42.